Fix Bug(#1): compile error, fix MySQL error and various fixes

+++ Authorization/Databases.cs
@@ -4,7 +4,7 @@ namespace Authorization {
         public static MySQL Auth;
 
         public static bool Init() {
-            Auth = new MySQL(Config.AUTH_DATABASE[0], int.Parse(Config.AUTH_DATABASE[1]), Config.AUTH_DATABASE[2], Config.AUTH_DATABASE[3], Config.AUTH_DATABASE[4]);
+            Auth = new MySQL(Config.AUTH_DATABASE[0], ushort.Parse(Config.AUTH_DATABASE[1]), Config.AUTH_DATABASE[2], Config.AUTH_DATABASE[3], Config.AUTH_DATABASE[4]);
             return Auth.Open();
         }
     }
