+++ Game/Databases.cs
@@ -4,7 +4,7 @@ namespace Game {
         public static MySQL Game;
 
         public static bool Init() {
-            Game = new MySQL(Config.GAME_DATABASE[0], int.Parse(Config.GAME_DATABASE[1]), Config.GAME_DATABASE[2], Config.GAME_DATABASE[3], Config.GAME_DATABASE[4]);
+            Game = new MySQL(Config.GAME_DATABASE[0], ushort.Parse(Config.GAME_DATABASE[1]), Config.GAME_DATABASE[2], Config.GAME_DATABASE[3], Config.GAME_DATABASE[4]);
             return Game.Open();
         }
     }
