+++ Game/Packets/LevelUp.cs
@@ -1,7 +1,10 @@
-﻿namespace Game.Packets {
-    class LevelUp : Core.Networking.OutPacket {
-        public LevelUp(Entities.User u, uint dinarEarned)
-            : base((ushort)Enums.Packets.LevelUp) {
+﻿namespace Game.Packets
+{
+    class LevelUp : Core.Networking.OutPacket
+    {
+        public LevelUp(Entities.User u, long dinarEarned)
+            : base((ushort)Enums.Packets.LevelUp)
+        {
             Append(u.RoomSlot);
             Append(0);
             Append(Core.LevelCalculator.GetLevelforExp(u.XP));
